requirements.txt
customtkinter>=5.2.2
darkdetect>=0.8.0
packaging>=26.0
matplotlib>=3.8.0
PyMuPDF>=1.24.0
requests>=2.31.0
beautifulsoup4>=4.12.0
pint>=0.23
sympy
numpy
Pillow
pytesseract
